pyproject.toml
[project]
name = "adk-tutorial"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk (>=1.2.1,<2.0.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "pydantic (>=2.11.5,<3.0.0)",
    "litellm (>=1.72.2,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
